Rename experiment.main => experiment.core
