+++ src/access_nri_intake/experiment/__init__.py
@@ -1,3 +1,3 @@
-from .main import use_datastore
+from .core import use_datastore
 
 __all__ = ["use_datastore"]
